fix(dependencies): add the ability to set $JAVA_HOME if $PATH contains Java bin directory
